chore: Move player_image ui system to own module
